+++ src/view/emoji.js
