+++ src/view/film-details.js
@@ -1,0 +1,97 @@
+export const createFilmDetailsTemplate =
+  (
+      title,
+      poster,
+      originalTitle,
+      rating,
+      director,
+      writers,
+      actors,
+      date,
+      duration,
+      country,
+      genres,
+      ageRating,
+      description,
+      isFavorite,
+      isInWatchList,
+      isWatched
+  ) => {
+    return (
+      `<section class="film-details">
+        <form class="film-details__inner" action="" method="get">
+          <div class="form-details__top-container">
+            <div class="film-details__close">
+              <button class="film-details__close-btn" type="button">close</button>
+            </div>
+            <div class="film-details__info-wrap">
+              <div class="film-details__poster">
+                <img class="film-details__poster-img" src="./images/posters/${poster}" alt="">
+                <p class="film-details__age">${ageRating}</p>
+              </div>
+              <div class="film-details__info">
+                <div class="film-details__info-head">
+                  <div class="film-details__title-wrap">
+                    <h3 class="film-details__title">${title}</h3>
+                    <p class="film-details__title-original">Original: ${originalTitle}</p>
+                  </div>
+                  <div class="film-details__rating">
+                    <p class="film-details__total-rating">${rating}</p>
+                  </div>
+                </div>
+                <table class="film-details__table">
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Director</td>
+                    <td class="film-details__cell">${director}</td>
+                  </tr>
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Writers</td>
+                    <td class="film-details__cell">${writers}</td>
+                  </tr>
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Actors</td>
+                    <td class="film-details__cell">${actors}</td>
+                  </tr>
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Release Date</td>
+                    <td class="film-details__cell">${date.day} ${date.month} ${date.year}</td>
+                  </tr>
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Runtime</td>
+                    <td class="film-details__cell">${duration}</td>
+                  </tr>
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Country</td>
+                    <td class="film-details__cell">${country}</td>
+                  </tr>
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Genres</td>
+                    <td class="film-details__cell">
+                      <span class="film-details__genre">${genres}</span>
+                      <span class="film-details__genre"></span>
+                      <span class="film-details__genre"></span>
+                    </td>
+                  </tr>
+                </table>
+
+                <p class="film-details__film-description">
+                 ${description}
+                </p>
+              </div>
+            </div>
+
+            <section class="film-details__controls">
+              <input type="checkbox" class="film-details__control-input visually-hidden" id="watchlist" name="watchlist" ${isInWatchList ? `checked` : ``}>
+              <label for="watchlist" class="film-details__control-label film-details__control-label--watchlist">Add to watchlist</label>
+
+              <input type="checkbox" class="film-details__control-input visually-hidden" id="watched" name="watched" ${isWatched ? `checked` : ``}>
+              <label for="watched" class="film-details__control-label film-details__control-label--watched" >Already watched</label>
+
+              <input type="checkbox" class="film-details__control-input visually-hidden" id="favorite" name="favorite" ${isFavorite ? `checked` : ``}>
+              <label for="favorite" class="film-details__control-label film-details__control-label--favorite">Add to favorites</label>
+            </section>
+          </div>
+        </form>
+      </section>`
+    );
+  };
