+++ src/view/top-rated-films-list.js
